🚚 renamings and tsconfig paths rework
