+++ apps/atom.io.fyi/tsconfig.json
@@ -1,4 +1,5 @@
 {
+	"extends": "../../tsconfig.json",
 	"compilerOptions": {
 		"target": "ES2017",
 		"lib": ["dom", "dom.iterable", "esnext"],
@@ -13,17 +14,13 @@
 		"isolatedModules": true,
 		"jsx": "preserve",
 		"incremental": true,
-		"plugins": [
-			{
-				"name": "next"
-			}
-		],
+		"plugins": [{	"name": "next"}],
 		"paths": {
-			"@/*": ["./src/app/*"],
-			"x/*": ["./src/exhibits-wrapped/*"],
-			"atom.io": ["../../packages/atom.io/src"],
-			"atom.io/*": ["../../packages/atom.io/*/src"],
-			"~/*": ["../../*"]
+			"~/*": ["./*"],
+			"src/*": ["~/apps/atom.io.fyi/src/*"],
+			"x/*": ["src/exhibits-wrapped/*"],
+			"atom.io": ["./packages/atom.io/src"],
+			"atom.io/*": ["./packages/atom.io/*/src"],
 		}
 	},
 	"include": [
@@ -31,7 +28,5 @@
 		"**/*.ts",
 		"**/*.tsx",
 		".next/types/**/*.ts",
-		"dist/types/**/*.ts"
 	],
-	"exclude": ["node_modules"]
 }
