+++ apps/tempest.games/src/components/PlayingCards.tsx
@@ -1,56 +1,56 @@
-import TwoOfSpades from "src/graphics/svg/playing-cards/2♠.svg?url"
-import TwoOfClubs from "src/graphics/svg/playing-cards/2♣.svg?url"
-import TwoOfHearts from "src/graphics/svg/playing-cards/2♥.svg?url"
-import TwoOfDiamonds from "src/graphics/svg/playing-cards/2♦.svg?url"
-import ThreeOfSpades from "src/graphics/svg/playing-cards/3♠.svg?url"
-import ThreeOfClubs from "src/graphics/svg/playing-cards/3♣.svg?url"
-import ThreeOfHearts from "src/graphics/svg/playing-cards/3♥.svg?url"
-import ThreeOfDiamonds from "src/graphics/svg/playing-cards/3♦.svg?url"
-import FourOfSpades from "src/graphics/svg/playing-cards/4♠.svg?url"
-import FourOfClubs from "src/graphics/svg/playing-cards/4♣.svg?url"
-import FourOfHearts from "src/graphics/svg/playing-cards/4♥.svg?url"
-import FourOfDiamonds from "src/graphics/svg/playing-cards/4♦.svg?url"
-import FiveOfSpades from "src/graphics/svg/playing-cards/5♠.svg?url"
-import FiveOfClubs from "src/graphics/svg/playing-cards/5♣.svg?url"
-import FiveOfHearts from "src/graphics/svg/playing-cards/5♥.svg?url"
-import FiveOfDiamonds from "src/graphics/svg/playing-cards/5♦.svg?url"
-import SixOfSpades from "src/graphics/svg/playing-cards/6♠.svg?url"
-import SixOfClubs from "src/graphics/svg/playing-cards/6♣.svg?url"
-import SixOfHearts from "src/graphics/svg/playing-cards/6♥.svg?url"
-import SixOfDiamonds from "src/graphics/svg/playing-cards/6♦.svg?url"
-import SevenOfSpades from "src/graphics/svg/playing-cards/7♠.svg?url"
-import SevenOfClubs from "src/graphics/svg/playing-cards/7♣.svg?url"
-import SevenOfHearts from "src/graphics/svg/playing-cards/7♥.svg?url"
-import SevenOfDiamonds from "src/graphics/svg/playing-cards/7♦.svg?url"
-import EightOfSpades from "src/graphics/svg/playing-cards/8♠.svg?url"
-import EightOfClubs from "src/graphics/svg/playing-cards/8♣.svg?url"
-import EightOfHearts from "src/graphics/svg/playing-cards/8♥.svg?url"
-import EightOfDiamonds from "src/graphics/svg/playing-cards/8♦.svg?url"
-import NineOfSpades from "src/graphics/svg/playing-cards/9♠.svg?url"
-import NineOfClubs from "src/graphics/svg/playing-cards/9♣.svg?url"
-import NineOfHearts from "src/graphics/svg/playing-cards/9♥.svg?url"
-import NineOfDiamonds from "src/graphics/svg/playing-cards/9♦.svg?url"
-import TenOfSpades from "src/graphics/svg/playing-cards/10♠.svg?url"
-import TenOfClubs from "src/graphics/svg/playing-cards/10♣.svg?url"
-import TenOfHearts from "src/graphics/svg/playing-cards/10♥.svg?url"
-import TenOfDiamonds from "src/graphics/svg/playing-cards/10♦.svg?url"
-import AceOfSpades from "src/graphics/svg/playing-cards/A♠.svg?url"
-import AceOfClubs from "src/graphics/svg/playing-cards/A♣.svg?url"
-import AceOfHearts from "src/graphics/svg/playing-cards/A♥.svg?url"
-import AceOfDiamonds from "src/graphics/svg/playing-cards/A♦.svg?url"
-import JackOfSpades from "src/graphics/svg/playing-cards/J♠.svg?url"
-import JackOfClubs from "src/graphics/svg/playing-cards/J♣.svg?url"
-import JackOfHearts from "src/graphics/svg/playing-cards/J♥.svg?url"
-import JackOfDiamonds from "src/graphics/svg/playing-cards/J♦.svg?url"
-import KingOfSpades from "src/graphics/svg/playing-cards/K♠.svg?url"
-import KingOfClubs from "src/graphics/svg/playing-cards/K♣.svg?url"
-import KingOfHearts from "src/graphics/svg/playing-cards/K♥.svg?url"
-import KingOfDiamonds from "src/graphics/svg/playing-cards/K♦.svg?url"
-import QueenOfSpades from "src/graphics/svg/playing-cards/Q♠.svg?url"
-import QueenOfClubs from "src/graphics/svg/playing-cards/Q♣.svg?url"
-import QueenOfHearts from "src/graphics/svg/playing-cards/Q♥.svg?url"
-import QueenOfDiamonds from "src/graphics/svg/playing-cards/Q♦.svg?url"
-import Back from "src/graphics/svg/playing-cards/back.svg?url"
+import TwoOfSpades from "tempest.games/graphics/svg/playing-cards/2♠.svg?url"
+import TwoOfClubs from "tempest.games/graphics/svg/playing-cards/2♣.svg?url"
+import TwoOfHearts from "tempest.games/graphics/svg/playing-cards/2♥.svg?url"
+import TwoOfDiamonds from "tempest.games/graphics/svg/playing-cards/2♦.svg?url"
+import ThreeOfSpades from "tempest.games/graphics/svg/playing-cards/3♠.svg?url"
+import ThreeOfClubs from "tempest.games/graphics/svg/playing-cards/3♣.svg?url"
+import ThreeOfHearts from "tempest.games/graphics/svg/playing-cards/3♥.svg?url"
+import ThreeOfDiamonds from "tempest.games/graphics/svg/playing-cards/3♦.svg?url"
+import FourOfSpades from "tempest.games/graphics/svg/playing-cards/4♠.svg?url"
+import FourOfClubs from "tempest.games/graphics/svg/playing-cards/4♣.svg?url"
+import FourOfHearts from "tempest.games/graphics/svg/playing-cards/4♥.svg?url"
+import FourOfDiamonds from "tempest.games/graphics/svg/playing-cards/4♦.svg?url"
+import FiveOfSpades from "tempest.games/graphics/svg/playing-cards/5♠.svg?url"
+import FiveOfClubs from "tempest.games/graphics/svg/playing-cards/5♣.svg?url"
+import FiveOfHearts from "tempest.games/graphics/svg/playing-cards/5♥.svg?url"
+import FiveOfDiamonds from "tempest.games/graphics/svg/playing-cards/5♦.svg?url"
+import SixOfSpades from "tempest.games/graphics/svg/playing-cards/6♠.svg?url"
+import SixOfClubs from "tempest.games/graphics/svg/playing-cards/6♣.svg?url"
+import SixOfHearts from "tempest.games/graphics/svg/playing-cards/6♥.svg?url"
+import SixOfDiamonds from "tempest.games/graphics/svg/playing-cards/6♦.svg?url"
+import SevenOfSpades from "tempest.games/graphics/svg/playing-cards/7♠.svg?url"
+import SevenOfClubs from "tempest.games/graphics/svg/playing-cards/7♣.svg?url"
+import SevenOfHearts from "tempest.games/graphics/svg/playing-cards/7♥.svg?url"
+import SevenOfDiamonds from "tempest.games/graphics/svg/playing-cards/7♦.svg?url"
+import EightOfSpades from "tempest.games/graphics/svg/playing-cards/8♠.svg?url"
+import EightOfClubs from "tempest.games/graphics/svg/playing-cards/8♣.svg?url"
+import EightOfHearts from "tempest.games/graphics/svg/playing-cards/8♥.svg?url"
+import EightOfDiamonds from "tempest.games/graphics/svg/playing-cards/8♦.svg?url"
+import NineOfSpades from "tempest.games/graphics/svg/playing-cards/9♠.svg?url"
+import NineOfClubs from "tempest.games/graphics/svg/playing-cards/9♣.svg?url"
+import NineOfHearts from "tempest.games/graphics/svg/playing-cards/9♥.svg?url"
+import NineOfDiamonds from "tempest.games/graphics/svg/playing-cards/9♦.svg?url"
+import TenOfSpades from "tempest.games/graphics/svg/playing-cards/10♠.svg?url"
+import TenOfClubs from "tempest.games/graphics/svg/playing-cards/10♣.svg?url"
+import TenOfHearts from "tempest.games/graphics/svg/playing-cards/10♥.svg?url"
+import TenOfDiamonds from "tempest.games/graphics/svg/playing-cards/10♦.svg?url"
+import AceOfSpades from "tempest.games/graphics/svg/playing-cards/A♠.svg?url"
+import AceOfClubs from "tempest.games/graphics/svg/playing-cards/A♣.svg?url"
+import AceOfHearts from "tempest.games/graphics/svg/playing-cards/A♥.svg?url"
+import AceOfDiamonds from "tempest.games/graphics/svg/playing-cards/A♦.svg?url"
+import JackOfSpades from "tempest.games/graphics/svg/playing-cards/J♠.svg?url"
+import JackOfClubs from "tempest.games/graphics/svg/playing-cards/J♣.svg?url"
+import JackOfHearts from "tempest.games/graphics/svg/playing-cards/J♥.svg?url"
+import JackOfDiamonds from "tempest.games/graphics/svg/playing-cards/J♦.svg?url"
+import KingOfSpades from "tempest.games/graphics/svg/playing-cards/K♠.svg?url"
+import KingOfClubs from "tempest.games/graphics/svg/playing-cards/K♣.svg?url"
+import KingOfHearts from "tempest.games/graphics/svg/playing-cards/K♥.svg?url"
+import KingOfDiamonds from "tempest.games/graphics/svg/playing-cards/K♦.svg?url"
+import QueenOfSpades from "tempest.games/graphics/svg/playing-cards/Q♠.svg?url"
+import QueenOfClubs from "tempest.games/graphics/svg/playing-cards/Q♣.svg?url"
+import QueenOfHearts from "tempest.games/graphics/svg/playing-cards/Q♥.svg?url"
+import QueenOfDiamonds from "tempest.games/graphics/svg/playing-cards/Q♦.svg?url"
+import Back from "tempest.games/graphics/svg/playing-cards/back.svg?url"
 
 export const PlayingCards: Record<string, SvgUrlData> = {
 	Back,
