+++ apps/tempest.games/src/game/__devtools__/Controls.tsx
@@ -4,18 +4,21 @@ import { useServerAction } from "atom.io/realtime-react"
 import { nanoid } from "nanoid"
 import type { FC } from "react"
 
-import { addHandTX, spawnClassicDeckTX } from "~/apps/node/lodge/src/store/game"
+import {
+	spawnClassicDeckTX,
+	spawnHandTX,
+} from "~/apps/node/lodge/src/store/game"
 
-import { button } from "src/components/<button>"
-import { myRoomState } from "src/services/store/my-room"
+import { button } from "tempest.games/components/<button>"
+import { myRoomState } from "tempest.games/services/store/my-room"
 
-import comic from "src/components/comic.module.scss"
+import comic from "tempest.games/components/comic.module.scss"
 import scss from "./Controls.module.scss"
 
 export const Controls: FC = () => {
 	const myId = useO(myIdState)
 	const myRoomId = useO(myRoomState)
-	const addHand = useServerAction(addHandTX)
+	const spawnHand = useServerAction(spawnHandTX)
 	const spawnClassicDeck = useServerAction(spawnClassicDeckTX)
 	return (
 		<span className={scss.class}>
@@ -24,7 +27,7 @@ export const Controls: FC = () => {
 					className={comic.class}
 					onClick={() => {
 						const groupId = nanoid()
-						addHand(myId, groupId)
+						spawnHand(myId, groupId)
 					}}
 				>
 					Add Hand
