+++ apps/tempest.games/src/game/labels/Count.tsx
@@ -1,4 +1,4 @@
-import { span } from "src/components/<span>"
+import { span } from "tempest.games/components/<span>"
 
 import scss from "./Count.module.scss"
 
@@ -6,7 +6,7 @@ export type CountProps = {
 	amount: number
 	minimal?: boolean
 }
-export function Count({ amount, minimal }: CountProps): JSX.Element {
+export function Count({ amount }: CountProps): JSX.Element {
 	return (
 		<span.diagon className={scss.class} data-css="count">
 			{amount}
