+++ apps/tempest.games/src/game/my-domain/MyDomain.tsx
@@ -1,8 +1,8 @@
 import { useO } from "atom.io/react"
 import * as React from "react"
 
-import { span } from "src/components/<span>"
-import { myHandsIndex } from "src/services/store/my-hands-index"
+import { span } from "tempest.games/components/<span>"
+import { myHandsIndex } from "tempest.games/services/store/my-hands-index"
 import { Hand } from "../game-pieces/Hand"
 import { Me } from "../my-domain/Me"
