+++ apps/tempest.games/src/game/public/Hearts.tsx
@@ -1,7 +1,7 @@
 import { useO } from "atom.io/react"
 
-import { publicDeckIndex } from "src/services/store/public-deck-index"
-import { publicTrickIndex } from "src/services/store/public-trick-index"
+import { publicDeckIndex } from "tempest.games/services/store/public-deck-index"
+import { publicTrickIndex } from "tempest.games/services/store/public-trick-index"
 import type { GameProps } from "../Game"
 import { Deck } from "../game-pieces/Deck"
 import { Trick } from "../game-pieces/Trick"
@@ -12,7 +12,7 @@ export function Hearts({ roomId }: GameProps): JSX.Element {
 	return (
 		<article className={scss.class}>
 			<section data-css="deck">
-				<GameDeck roomId={roomId} />
+				<GameDeck />
 			</section>
 			<section data-css="trick">
 				<CurrentTrick roomId={roomId} />
@@ -20,7 +20,7 @@ export function Hearts({ roomId }: GameProps): JSX.Element {
 		</article>
 	)
 }
-function GameDeck({ roomId }: GameProps): JSX.Element | null {
+function GameDeck(): JSX.Element | null {
 	const deckId = useO(publicDeckIndex)[0]
 	return deckId ? <Deck id={deckId} /> : null
 }
