+++ packages/rel8/types/src/json.ts
@@ -1,4 +1,4 @@
 import type * as Json from "~/packages/anvl/src/json/json"
-import { primitive } from "~/packages/anvl/src/primitive"
+import type { primitive } from "~/packages/anvl/src/primitive"
 
-export { primitive, Json }
+export type { primitive, Json }
